Fix: Refactored trivia handler to endless trivia handler
